requirements.txt
web3==5.19.0
eth-abi==2.1.1
eth-utils==1.9.5
python-dateutil==2.7.0
click==8.0.1
ethereum-dasm==0.1.4
base58
requests
pytest~=4.3.0
timeout-decorator==0.4.1
google-cloud-pubsub==0.39.1
sqlalchemy==1.3.13
pg8000==1.13.2
pymongo==3.11.4
pyetherchain==0.3.2
python-dotenv==0.19.2